pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "General-purpose LLMOps platform with pluggable scoring, tuning, and observability"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # LLM providers
    "anthropic>=0.42.0",
    "google-genai>=1.0.0",
    # Observability
    "langfuse>=2.51.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Redis
    "redis[hiredis]>=5.2.0",
    # HTTP client
    "httpx>=0.28.0",
    # Config
    "pydantic-settings>=2.6.0",
    # Prompt rendering
    "jinja2>=3.1.0",
    # Structured logging
    "structlog>=24.0.0",
    # Prometheus metrics
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmops"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "integration: marks tests as integration tests (require running DB + Redis)",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
